unidades/06_acceso_a_bases_de_datos/ninja_trainings/02_bark/bark_done/bark/requirements.txt
gunicorn~=21.2.0
mysql-connector-python~=8.2.0
multipart~=0.2.4
tzdata~=2023.3
